rename back to cli
