+++ src/cli-command/createCliCommand.ts
@@ -1,5 +1,5 @@
 import { RecursivePartial } from 'type-plus';
-import { buildContext, CliContext } from '../basic-cli';
+import { buildContext, CliContext } from '../cli';
 import { log } from '../log';
 import { CliCommand, CliCommandInstance } from './CliCommand';
 import { MultipleArgumentNotLastEntry, OptionNameNotUnique } from './errors';
